+++ js/src/widget.js
@@ -59,8 +59,8 @@ function chosen_on_array_add(inst, context) {
 }
 
 $(function() {
-    if (yafowil_array === undefined) {
+    if (window.yafowil_array === undefined) {
         return;
     }
-    yafowil_array.on_array_event('on_add', chosen_on_array_add);
+    window.yafowil_array.on_array_event('on_add', chosen_on_array_add);
 });