+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 2.0 (unreleased)
 ----------------
 
+- Extend JS by ``chosen_on_array_add`` function to enable usage in
+  ``yafowil.widget.array``.
+  [lenadax]
+
 - Rewrite JavaScript using ES6.
   [rnix]